+++ CHANGELOG.md
@@ -24,7 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Adds a file format for external graphs, defined as a schema in ISL (Ion Schema Language).
+- Adds a file format for external graphs, defined as a schema in ISL (Ion Schema Language), 
+  as well as an in-memory graph data model and a reader for loading external graphs into it.
 
 ### Changed